refactor(mtfdigitizer): decompose ridge.py into a staged package (#1414)
Split the 2104-line pipeline/ridge.py into a pipeline/ridge/ package by
data-flow stage (ADR-083):

- foundation.py — mask chrome-strip, per-column ridge extraction, track
  clustering/merge/select, rasterize/densify/coincidence (+ constants)
- dp.py — per-column Viterbi DP passes and crossing detection/swap
- hue.py — ridge_tracks_for_hue_freq_split (FREQUENCY_PER_HUE_RIDGE)
- fields.py — multifreq band assignment + ridge_tracks_to_fields_multifreq
  / ridge_tracks_to_fields (RIDGE_TRACKING)
- __init__.py — re-exports the public entry points and the private
  helpers test_ridge.py imports (in __all__)

The dependency graph is acyclic (foundation <- dp <- hue; foundation <-
fields). The public import path is unchanged, so dispatch.py and
test_ridge.py are not modified — the unchanged suite is the regression
oracle.

Output-neutral: calibrate byte-identical (zero GT delta), full pytest
511 passed, pyflakes clean. git renders this as delete + five adds — a
split is not a rename; no logic changed. ridge_tracks_to_fields (tested,
production-unused wrapper) is preserved.

Sub-task 3 of 3 in #1414. Closes #1414.
